big push

+++ eval_array.sh
@@ -1,42 +1,31 @@
 #!/bin/bash
-#SBATCH --job-name=eval_hf
+#SBATCH --job-name=eval_arr
 #SBATCH --output=logs/%u/%x_%j.out
 #SBATCH --error=logs/%u/%x_%j.err
 #SBATCH --ntasks=1
 #SBATCH --tmp=64G
-#SBATCH --mem-per-cpu=32G
+#SBATCH --mem-per-cpu=8G
 #SBATCH --gpus-per-node=1
 #SBATCH --nodes=1
 #SBATCH --time=24:00:00
 #SBATCH --account=es_sachan
 #SBATCH --mail-type=END,FAIL
-#SBATCH --gres=gpumem:32g
+
+#SBATCH --gres=gpumem:64g
 
 # for 12 gemma -> gpu 32 cpu 64 batch 1 worked -> 16 with batchsize 4 didnt work -> 32g batch 4 tbd 32 cpu didnt work ->32 with fast mode batch 1
 
-# MODEL=${1:-"openai/gpt-4-1"}
-# MODEL=${1:-"openai/o4-mini"}
-
-
-# MODEL=${1:-"huggingface/Qwen2-5-VL-3B-Instruct"}
-# MODEL=${1:-"huggingface/Qwen2-5-VL-7B-Instruct"}
-
-
-MODEL=${1:-"huggingface/gemma-3-4b-it"}
-# MODEL=${1:-"huggingface/gemma-3-12b-it"}
-# MODEL=${1:-"huggingface/gemma-3-27b-it"}
-
-# MODEL=${1:-"huggingface/Llama-4-Scout-17B-16E-Instruct"}
-
-# MODEL=${1:-"google/gemini-2.5-flash-preview-04-17"}
-# MODEL=${1:-"google/gemini-2.5-pro-preview-05-06"}
-# MODEL=${1:-"xai/grok-2-vision"}
+models=(
+    "huggingface/gemma-3-12b-it"
+    "huggingface/gemma-3-27b-it"
+)
+MODEL=${models[$SLURM_ARRAY_TASK_ID]}
 
 # Load the necessary modules
 module load stack/2024-06 gcc/12.2.0 python/3.11.6 cuda/11.3.1 eth_proxy
 
-## Delete the venv
-if [ -d "$SCRATCH/pmlr/$MODEL/.venv" ]; then
+## Delete the venv if flag is set
+if [ "$1" = "clear" ] && [ -d "$SCRATCH/pmlr/$MODEL/.venv" ]; then
   rm -rf "$SCRATCH/pmlr/$MODEL/.venv"
   echo "Cleared existing virtual environment at $SCRATCH/pmlr/$MODEL/.venv"
 fi
@@ -56,25 +45,15 @@ pip install -r requirements.txt --quiet
 
 # Model paths
 export HF_HOME="$SCRATCH/pmlr/$MODEL/cache"
+export TRANSFORMERS_CACHE="$SCRATCH/pmlr/$MODEL/cache"
 
 echo "$USER starting Benchmarking job for $MODEL"
 echo "Job started at $(date)"
 
 python src/run_eval.py \
   --config $MODEL \
-  --datasets "Sample" 
-  # --datasets "GeckoNum, PixMo_Count"
-  #--datasets "FSC-147" \ 
-  #--datasets "PixMo_Count" \
-  #--datasets "GeckoNum" 
-  # --datasets "Missing"
-  # --datasets "Sample" \
-  # --datasets "FSC-147" \   
-  
-  # --batch_size 4
-  # --batch_size 1
+  --datasets "TallyQA"
   # --datasets "FSC-147, TallyQA" \
-  # --batch_size -1 \
 
 echo "Job completed at $(date)"
